+++ bin/samples/sample-static.js
@@ -6,8 +6,11 @@
  * @method get
  */
 /* eslint-disable */
+const faker = require('http-request-mock/plugin/faker.js');
 module.exports = {
   ret: 0,
   msg: 'ok',
+  ip: faker.ip(),
+  url: faker.url(),
   data: 'The request will be resolved after 1000 milliseconds.',
 };